+++ src/targets/npm.ts
@@ -7,6 +7,7 @@ import {
 	ruleCompile,
 	extractPackageJson,
 	extractGitignore,
+	type InternalRules,
 } from "../patterns/index.js"
 import { unixify } from "../unixify.js"
 import { npmManifestParse, type PackageJson } from "./npmManifest.js"
@@ -30,62 +31,83 @@ export function makeNPM(): Target {
 		},
 	]
 
+	const bundledInclude: Rule = {
+		compiled: [],
+		excludes: false,
+		pattern: [], // filled within init
+	}
 	const internalInclude: Rule = {
 		compiled: [],
 		excludes: false,
 		pattern: [], // filled within init
 	}
 
-	const internal: Rule[] = [
-		internalInclude,
-		ruleCompile({
-			compiled: null,
-			excludes: true,
-			pattern: [
-				// https://github.com/npm/npm-packlist/blob/main/lib/index.js#L16
-				".npmignore",
-				".gitignore",
-				".git",
-				".svn",
-				".hg",
-				"CVS",
-				"/.lock-wscript",
-				"/.wafpickle-*",
-				"/build/config.gypi",
-				"npm-debug.log",
-				".npmrc",
-				".*.swp",
-				".DS_Store",
-				"._*",
-				"*.orig",
-				"/archived-packages/**",
+	const internal: InternalRules = {
+		after: [
+			ruleCompile(
+				{
+					compiled: null,
+					excludes: true,
+					pattern: [".npmignore", ".gitignore"],
+				},
+				{ nocase: true },
+			),
+		],
+		before: [
+			internalInclude,
+			bundledInclude,
+			ruleCompile(
+				{
+					compiled: null,
+					excludes: true,
+					pattern: [
+						// https://github.com/npm/npm-packlist/blob/main/lib/index.js#L16
+						".git",
+						".svn",
+						".hg",
+						"CVS",
+						"/.lock-wscript",
+						"/.wafpickle-*",
+						"/build/config.gypi",
+						"npm-debug.log",
+						".npmrc",
+						".*.swp",
+						".DS_Store",
+						"._*",
+						"*.orig",
+						"/archived-packages/**",
 
-				// https://github.com/npm/npm-packlist/blob/main/lib/index.js#L294
-				"/node_modules",
-				"/package-lock.json",
-				"/yarn.lock",
-				"/pnpm-lock.yaml",
-				"/bun.lockb",
-			],
-		}),
-		ruleCompile({
-			compiled: null,
-			excludes: false,
-			pattern: [
-				// https://github.com/npm/npm-packlist/blob/main/lib/index.js#L287
-				"bin",
-				"package.json",
-				"README",
-				"COPYING",
-				"LICENSE",
-				"LICENCE",
-				"README.*",
-				"COPYING.*",
-				"LICENSE.*",
-				"LICENCE.*",
-			],
-		}),
-	]
+						// https://github.com/npm/npm-packlist/blob/main/lib/index.js#L294
+						"/node_modules",
+						"/package-lock.json",
+						"/yarn.lock",
+						"/pnpm-lock.yaml",
+						"/bun.lockb",
+					],
+				},
+				{ nocase: true },
+			),
+			ruleCompile(
+				{
+					compiled: null,
+					excludes: false,
+					pattern: [
+						// https://github.com/npm/npm-packlist/blob/main/lib/index.js#L287
+						"/package.json",
+						"README",
+						"COPYING",
+						"LICENSE",
+						"LICENCE",
+						"README.*",
+						"COPYING.*",
+						"LICENSE.*",
+						"LICENCE.*",
+					],
+				},
+				{ nocase: true },
+			),
+		],
+	}
 
 	return <Target>{
 		extractors,